book-api/src/requirements.txt
Flask==0.12
mysql-connector
